+++ packages/desktop/app/components/Surveys/SurveyScreen.js
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { runCalculations } from 'shared/utils/calculations';
 import styled from 'styled-components';
 import { checkVisibility } from '../../utils';
 import { FormGrid } from '../FormGrid';
@@ -10,15 +11,29 @@ const StyledButtonRow = styled(ButtonRow)`
   margin-top: 24px;
 `;
 
+const useCalculatedFormValues = (components, values, setFieldValue) => {
+  useEffect(() => {
+    // recalculate dynamic fields
+    const calculatedValues = runCalculations(components, values);
+    // write values that have changed back into answers
+    Object.entries(calculatedValues)
+      .filter(([k, v]) => values[k] !== v)
+      .map(([k, v]) => setFieldValue(k, v));
+  }, [components, values, setFieldValue]);
+};
+
 export const SurveyScreen = ({
   components,
   values = {},
+  setFieldValue,
   onStepForward,
   onStepBack,
   submitButton,
   patient,
   cols = 1,
 }) => {
+  useCalculatedFormValues(components, values, setFieldValue);
+
   const questionElements = components
     .filter(c => checkVisibility(c, values, components))
     .map(c => <SurveyQuestion component={c} patient={patient} key={c.id} />);
