+++ packages/desktop/app/forms/VitalsForm.js
@@ -7,7 +7,7 @@ import { getCurrentDateTimeString } from 'shared/utils/dateTime';
 import { ModalLoader, ConfirmCancelRow, Form } from '../components';
 import { SurveyScreen } from '../components/Surveys';
 import { useVitalsSurvey } from '../api/queries';
-import { getValidationSchema } from '../utils';
+import { getFormInitialValues, getValidationSchema } from '../utils';
 
 const ErrorMessage = () => {
   return (
@@ -20,7 +20,7 @@ const ErrorMessage = () => {
   );
 };
 
-export const VitalsForm = React.memo(({ patient, onSubmit, onClose, editedObject }) => {
+export const VitalsForm = React.memo(({ patient, onSubmit, onClose }) => {
   const { data: vitalsSurvey, isLoading, isError } = useVitalsSurvey();
   const validationSchema = useMemo(() => getValidationSchema(vitalsSurvey), [vitalsSurvey]);
 
@@ -42,7 +42,7 @@ export const VitalsForm = React.memo(({ patient, onSubmit, onClose, editedObject
       validationSchema={validationSchema}
       initialValues={{
         [VITALS_DATA_ELEMENT_IDS.dateRecorded]: getCurrentDateTimeString(),
-        ...editedObject,
+        ...getFormInitialValues(vitalsSurvey.components, patient),
       }}
       validate={({ [VITALS_DATA_ELEMENT_IDS.dateRecorded]: date, ...values }) => {
         const errors = {};
@@ -54,12 +54,14 @@ export const VitalsForm = React.memo(({ patient, onSubmit, onClose, editedObject
 
         return errors;
       }}
-      render={({ submitForm }) => {
+      render={({ submitForm, values, setFieldValue }) => {
         return (
           <SurveyScreen
             components={vitalsSurvey.components}
             patient={patient}
             cols={2}
+            values={values}
+            setFieldValue={setFieldValue}
             submitButton={
               <ConfirmCancelRow confirmText="Record" onConfirm={submitForm} onCancel={onClose} />
             }
@@ -74,9 +76,4 @@ VitalsForm.propTypes = {
   onSubmit: PropTypes.func.isRequired,
   onClose: PropTypes.func.isRequired,
   patient: PropTypes.object.isRequired,
-  editedObject: PropTypes.shape({}),
-};
-
-VitalsForm.defaultProps = {
-  editedObject: null,
 };
